+++ web/crud-api/resources/views/roles/list.blade.php
@@ -5,7 +5,11 @@
             <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
                 {{ __('Roles') }}
             </h2>
-            <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">Create</a>
+
+            @can('create roles')
+                <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">Create</a>
+            @endcan
+
         </div>
     </x-slot>
 
@@ -41,12 +45,17 @@
                                     {{-- \Carbon\Carbon::parse => format datetime --}}
                                     {{ \Carbon\Carbon::parse($role->created_at)->format('d M, Y') }}</td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('roles.edit', $role->id) }}"
-                                        class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600 ">Edit</a>
+                                    @can('edit roles')
+                                        <a href="{{ route('roles.edit', $role->id) }}"
+                                            class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600 ">Edit</a>
+                                    @endcan
 
-                                    {{-- onclick="deletePermission({{ $permission->id }})"     --}}
-                                    <a href="javascript:void(0);" data-id="{{ $role->id }}"
-                                        class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600 deleteBtn">Delete</a>
+                                    @can('delete roles')
+                                        {{-- onclick="deletePermission({{ $permission->id }})"     --}}
+                                        <a href="javascript:void(0);" data-id="{{ $role->id }}"
+                                            class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600 deleteBtn">Delete</a>
+                                    @endcan
+
                                 </td>
                             </tr>
                         @endforeach
